Moved "formatDate" multiple files to "Date.js"
