+++ src/Components/Comment/Comment.js
@@ -4,16 +4,9 @@ import {Link} from "react-router-dom";
 import './Comment.css';
 import Vote from "../Vote/Vote";
 
+import {formatDate} from "../../Utility/Date";
+
 function Comment(props) {
-    function formatDate(dateParam){
-        const date = new Date(dateParam);
-        const day = ("0" + date.getDate()).slice(-2);
-        const month = ("0" + (date.getMonth()+1)).slice(-2);
-        const year = date.getFullYear();
-        const hours = ("0" + date.getHours()).slice(-2);
-        const minutes = ("0" + date.getMinutes()).slice(-2);
-        return day +'.' + month +'.' + year + ' ' + hours +':' + minutes;
-    }
 
     return (
         <div className="Comment">
