+++ src/Components/PostPage/PostPage.js
@@ -7,6 +7,8 @@ import Comment from "../Comment/Comment";
 import AddComment from "../AddComment/AddComment";
 import Vote from "../Vote/Vote";
 
+import {formatDate} from "../../Utility/Date";
+
 function PostPage(props) {
     const [post, setPost] = useState([]);
     const [username, setUsername] = useState('');
@@ -24,16 +26,6 @@ function PostPage(props) {
         setPost(response.data);
         setUsername(response.data.postAuthor.username);
     }, []);
-
-    function formatDate(dateParam){
-        const date = new Date(dateParam);
-        const day = ("0" + date.getDate()).slice(-2);
-        const month = ("0" + (date.getMonth()+1)).slice(-2);
-        const year = date.getFullYear();
-        const hours = ("0" + date.getHours()).slice(-2);
-        const minutes = ("0" + date.getMinutes()).slice(-2);
-        return day +'.' + month +'.' + year + ' ' + hours +':' + minutes;
-    }
 
     return (
         <div className="PostPage">
